refactor: split local runtime monolith
